move to use service rather than the full compose file
