+++ src/convert/convert_to_containerapps.rs
@@ -1,4 +1,4 @@
-use crate::compose::Compose;
+use crate::compose::Service;
 use crate::containerapps::ContainerAppConfig;
 use anyhow::Result;
 use std::collections::HashMap;
@@ -6,8 +6,8 @@ use std::collections::HashMap;
 use super::get_properties;
 
 pub fn convert_to_containerapps(
-    compose_file: Compose,
-    cli_values: HashMap<&str, String>,
+    service: Service,
+    cli_values: &HashMap<&str, String>,
 ) -> Result<ContainerAppConfig> {
     let config = ContainerAppConfig {
         kind: "containerapp".to_string(),
@@ -16,8 +16,55 @@ pub fn convert_to_containerapps(
         location: cli_values["location"].to_owned(),
         resource_type: "Microsoft.Web/containerApps".to_string(),
         tags: None,
-        properties: get_properties(cli_values["kubeEnvironmentId"].to_owned(), &compose_file)?,
+        properties: get_properties(cli_values["kubeEnvironmentId"].to_owned(), &service)?,
     };
 
     Ok(config)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::convert::convert_to_containerapps;
+    use crate::convert::tests::{
+        get_sample_cli_args, get_service_from_docker_compose_file, read_containerapps_file,
+    };
+
+    #[test]
+    fn conversion_sets_name_from_cli() {
+        let compose_config = get_service_from_docker_compose_file();
+        let cli_args = get_sample_cli_args();
+        let new_containerapps_config = convert_to_containerapps(compose_config, &cli_args).unwrap();
+
+        let reference_containerapps_config = read_containerapps_file();
+        assert_eq!(
+            new_containerapps_config.name,
+            reference_containerapps_config.name
+        );
+    }
+
+    #[test]
+    fn conversion_sets_resource_group_from_cli() {
+        let compose_config = get_service_from_docker_compose_file();
+        let cli_args = get_sample_cli_args();
+        let new_containerapps_config = convert_to_containerapps(compose_config, &cli_args).unwrap();
+
+        let reference_containerapps_config = read_containerapps_file();
+        assert_eq!(
+            new_containerapps_config.resource_group,
+            reference_containerapps_config.resource_group
+        );
+    }
+
+    #[test]
+    fn conversion_sets_location_from_cli() {
+        let compose_config = get_service_from_docker_compose_file();
+        let cli_args = get_sample_cli_args();
+        let new_containerapps_config = convert_to_containerapps(compose_config, &cli_args).unwrap();
+
+        let reference_containerapps_config = read_containerapps_file();
+        assert_eq!(
+            new_containerapps_config.location,
+            reference_containerapps_config.location
+        );
+    }
 }
