+++ src/convert/get_container_from_service.rs
@@ -2,18 +2,18 @@ use crate::compose::Service;
 use crate::containerapps::{Container, EnvironmentConfiguration};
 use anyhow::Result;
 
-pub fn get_container_from_service(service: Service) -> Result<Container> {
+pub fn get_container_from_service(service: &Service) -> Result<Container> {
     let mut container = Container::default();
-    if let Some(image) = service.image {
+    if let Some(image) = &service.image {
         container.image = image.value()?.to_string();
     }
 
-    if let Some(name) = service.container_name {
+    if let Some(name) = &service.container_name {
         container.name = Some(name.value()?.to_string());
     }
 
     if !service.environment.is_empty() {
-        for (key, wrapped_value) in service.environment.into_iter() {
+        for (key, wrapped_value) in service.environment.clone().into_iter() {
             let new_value = match wrapped_value.value() {
                 Ok(v) => Some(v.to_string()),
                 _ => None,
@@ -27,4 +27,94 @@ pub fn get_container_from_service(service: Service) -> Result<Container> {
         }
     }
     Ok(container)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::convert::convert_to_containerapps;
+    use crate::convert::tests::*;
+
+    #[test]
+    fn conversion_sets_properties_template_containers_image() {
+        let compose_config = get_service_from_docker_compose_file();
+        let cli_args = get_sample_cli_args();
+        let new_containerapps_config = convert_to_containerapps(compose_config, &cli_args).unwrap();
+
+        let reference_containerapps_config = read_containerapps_file();
+        let new_ghost_container = new_containerapps_config
+            .properties
+            .template
+            .containers
+            .first()
+            .unwrap();
+        let reference_ghost_container = reference_containerapps_config
+            .properties
+            .template
+            .containers
+            .first()
+            .unwrap();
+
+        assert_eq!(new_ghost_container.image, reference_ghost_container.image);
+    }
+
+    #[test]
+    fn conversion_sets_properties_template_containers_name() {
+        let compose_config = get_service_from_docker_compose_file();
+        let cli_args = get_sample_cli_args();
+        let new_containerapps_config = convert_to_containerapps(compose_config, &cli_args).unwrap();
+
+        let reference_containerapps_config = read_containerapps_file();
+        let new_ghost_container = new_containerapps_config
+            .properties
+            .template
+            .containers
+            .first()
+            .unwrap();
+        let reference_ghost_container = reference_containerapps_config
+            .properties
+            .template
+            .containers
+            .first()
+            .unwrap();
+        assert_eq!(
+            new_ghost_container.name.clone().unwrap(),
+            reference_ghost_container.name.clone().unwrap()
+        );
+    }
+
+    #[test]
+    fn conversion_sets_properties_template_containers_environment() {
+        let compose_config = get_service_from_docker_compose_file();
+        let cli_args = get_sample_cli_args();
+        let new_containerapps_config = convert_to_containerapps(compose_config, &cli_args).unwrap();
+
+        let reference_containerapps_config = read_containerapps_file();
+        let new_ghost_container_env = new_containerapps_config
+            .properties
+            .template
+            .containers
+            .first()
+            .unwrap()
+            .env
+            .first()
+            .unwrap();
+        let reference_ghost_container_env = reference_containerapps_config
+            .properties
+            .template
+            .containers
+            .first()
+            .unwrap()
+            .env
+            .first()
+            .unwrap();
+
+        assert_eq!(
+            new_ghost_container_env.name,
+            reference_ghost_container_env.name
+        );
+        assert_eq!(
+            new_ghost_container_env.value.clone().unwrap(),
+            reference_ghost_container_env.value.clone().unwrap()
+        );
+    }
 }
